requirements.txt
absl-py==2.0.0
ale-py==0.8.1
aniso8601==9.0.1
anyio==3.6.2
appnope==0.1.3
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
boto3==1.9.253
botocore==1.12.253
box2d-py==2.3.5
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==7.1.2
cloudpickle==2.2.1
cmd2==0.9.25
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cryptography==41.0.2
cycler==0.11.0
Cython==0.29.36
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.4
executing==1.2.0
Faker==1.0.8
Farama-Notifications==0.0.4
fasteners==0.19
fastjsonschema==2.16.2
Flask==1.1.4
Flask-Login==0.4.1
Flask-RESTful==0.3.10
fonttools==4.38.0
fqdn==1.5.1
future==0.17.1
fuzzywuzzy==0.17.0
gevent==23.7.0
gitdb==4.0.10
GitPython==3.1.32
glfw==2.6.2
greenlet==2.0.2
gunicorn==20.0.4
gym==0.26.2
gym-notices==0.0.8
gym-recording==0.0.1
gymnasium==0.29.1
idna==2.8
imageio==2.31.4
imageio-ffmpeg==0.4.9
importlib-metadata==6.0.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==1.1.0
jax==0.4.16
jaxlib==0.4.16
jedi==0.18.2
Jinja2==2.11.3
jmespath==0.10.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-cache==0.5.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lz4==4.3.2
MarkupSafe==2.0.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.4
ml-dtypes==0.3.1
moviepy==1.0.3
mpmath==1.2.1
mujoco==2.3.7
mujoco-py==2.1.2.14
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
pluggy==0.13.1
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.36
PsiTurk==3.3.1
psutil==5.6.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
PyBrain==0.3
pycparser==2.21
pygame==2.5.2
Pygments==2.14.0
PyOpenGL==3.1.7
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
pytest==6.1.2
python-dateutil==2.8.2
python-dotenv==0.12.0
python-json-logger==2.0.4
pytz==2019.3
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.22.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
s3transfer==0.2.1
scipy==1.10.0
Send2Trash==1.8.0
setproctitle==1.1.10
Shimmy==0.2.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.24
stack-data==0.6.2
swig==4.1.1
sympy==1.11.1
tabulate==0.9.0
terminado==0.17.1
text-unidecode==1.2
tinycss2==1.2.1
toml==0.10.2
torch==1.13.1
tornado==6.2
tqdm==4.66.1
traitlets==5.8.1
typing_extensions==4.4.0
tzlocal==5.0.1
ua-parser==0.18.0
uri-template==1.2.0
urllib3==1.25.11
user-agents==2.2.0
-e git+https://github.com/rubenvereecken/py-vgdl@e1d491c658c62f01bdfe391f0f7815320caae5b0#egg=vgdl
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==1.0.1
widgetsnbextension==4.0.5
zipp==3.11.0
zope.event==5.0
zope.interface==6.0
